needs fixing & review

+++ src/main/flight/gps_rescue.h
@@ -70,7 +70,7 @@ typedef struct {
 } rescueState_s;
 
 extern bool canUseGPSHeading;
-extern int16_t gpsRescueAngle[ANGLE_INDEX_COUNT]; //in decidegrees
+extern int32_t gpsRescueAngle[ANGLE_INDEX_COUNT]; //in decidegrees
 extern rescueState_s rescueState;
 
 void updateGPSRescueState(void);
